+++ src/main/java/fantastic/faniverse/community/controller/CommentController.java
@@ -6,6 +6,8 @@ import fantastic.faniverse.community.controller.dto.CommentUpdateForm;
 import fantastic.faniverse.community.service.CommentService;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -18,6 +20,7 @@ import java.util.stream.Collectors;
 @RequestMapping("/comment")
 public class CommentController {
     private final CommentService commentService;
+    private static final Logger logger = LoggerFactory.getLogger(CommentController.class);
 
     @PostMapping("/create/{postId}")
     public ResponseEntity<Object> creatComment(
@@ -26,7 +29,14 @@ public class CommentController {
             HttpSession session
     ) {
         Long userId = (Long) session.getAttribute("userId");
+
+        if (userId == null) {
+            logger.error("User ID is null. Make sure the user is logged in.");
+            return ResponseEntity.status(401).body("User not logged in");
+        }
+
         commentService.createComment(request.toEntity(), postId, userId);
+        logger.info("User ID: " + userId + " created a comment on Post ID: " + postId);
         return ResponseEntity.ok().build();
     }
 
@@ -49,7 +59,14 @@ public class CommentController {
             HttpSession session
     ) {
         Long userId = (Long) session.getAttribute("userId");
+
+        if (userId == null) {
+            logger.error("User ID is null. Make sure the user is logged in.");
+            return ResponseEntity.status(401).body("User not logged in");
+        }
+
         commentService.updateComment(updateForm.toEntity(), commentId, userId);
+        logger.info("User ID: " + userId + " updated Comment ID: " + commentId);
         return ResponseEntity.ok().build();
     }
 
@@ -59,7 +76,14 @@ public class CommentController {
             HttpSession session
     ) {
         Long userId = (Long) session.getAttribute("userId");
+
+        if (userId == null) {
+            logger.error("User ID is null. Make sure the user is logged in.");
+            return ResponseEntity.status(401).body("User not logged in");
+        }
+
         commentService.delete(commentId, userId);
+        logger.info("User ID: " + userId + " deleted Comment ID: " + commentId);
         return ResponseEntity.ok().build();
     }
 }
